# lib/api-zod/src/generated/types/composition.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ingredient } from "./ingredient";

export interface Composition {
  calories: number;
  sugarGrams: number;
  caffeineMg: number;
  additives: string[];
  ingredients: Ingredient[];
}