# lib/api-zod/src/generated/types/ingredient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IngredientHealthRole } from "./ingredientHealthRole";
import type { IngredientRiskLevel } from "./ingredientRiskLevel";

export interface Ingredient {
  name: string;
  function: string;
  healthRole: IngredientHealthRole;
  riskLevel: IngredientRiskLevel;
}