# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Composition } from "./composition";
import type { LongTermImpact } from "./longTermImpact";
import type { MediumTermImpact } from "./mediumTermImpact";
import type { ScanResultGlycemicImpact } from "./scanResultGlycemicImpact";
import type { ScanResultStatus } from "./scanResultStatus";
import type { ShortTermImpact } from "./shortTermImpact";

export interface ScanResult {
  id: string;
  detectedProduct: string;
  brand?: string | null;
  category: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  impactScore: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  hydrationLevel: number;
  glycemicImpact: ScanResultGlycemicImpact;
  status: ScanResultStatus;
  aiInsight: string;
  viralStatement: string;
  shortTermImpact: ShortTermImpact;
  mediumTermImpact: MediumTermImpact;
  longTermImpact: LongTermImpact;
  composition: Composition;
}